# backend/api/openapi.yaml
openapi: "3.1.0"
info:
  title: SpendWise API
  description: Collaborative household financial ledger API
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/workspaces:
    post:
      operationId: createWorkspace
      tags: [Workspaces]
      summary: Create a new workspace
      description: Creates a workspace and adds the authenticated user as owner.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkspaceRequest"
      responses:
        "201":
          description: Workspace created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

    get:
      operationId: listWorkspaces
      tags: [Workspaces]
      summary: List workspaces the user is a member of
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of workspaces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkspaceResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/workspaces/{workspaceId}:
    get:
      operationId: getWorkspace
      tags: [Workspaces]
      summary: Get workspace by ID
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "200":
          description: Workspace details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      operationId: updateWorkspace
      tags: [Workspaces]
      summary: Update workspace (editor+ required)
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorkspaceRequest"
      responses:
        "200":
          description: Workspace updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      operationId: deleteWorkspace
      tags: [Workspaces]
      summary: Delete workspace (owner only)
      description: Permanently deletes the workspace and all associated data (members, categories, transactions, funding).
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "204":
          description: Workspace deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ──────────────────────────────────────────
  # Workspace Members
  # ──────────────────────────────────────────
  /api/v1/workspaces/{workspaceId}/members:
    post:
      operationId: addMember
      tags: [Workspaces]
      summary: Add a member to workspace (owner only)
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddMemberRequest"
      responses:
        "201":
          description: Member added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/workspaces/{workspaceId}/members/{userId}:
    delete:
      operationId: removeMember
      tags: [Workspaces]
      summary: Remove a member from workspace (owner only)
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/UserId"
      responses:
        "204":
          description: Member removed
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ──────────────────────────────────────────
  # Categories
  # ──────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT obtained from `/api/v1/auth/{provider}/callback`
  parameters:
...
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Insufficient permission
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "workspace not found"
      required: [error]
...
    CreateWorkspaceRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          example: Home
        description:
          type: string
          minLength: 1
          example: Household expenses
      required: [name, description]
    UpdateWorkspaceRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
      required: [name, description]
    WorkspaceResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        owner_id:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, name, description, owner_id, created_at, updated_at]
    AddMemberRequest:
      type: object
      properties:
        user_id:
          type: integer
          format: int64
        role:
          type: string
          enum: [editor, viewer]
      required: [user_id, role]
    MemberResponse:
      type: object
      properties:
        workspace_id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        role:
          type: string
          enum: [owner, editor, viewer]
        created_at:
          type: string
          format: date-time
      required: [workspace_id, user_id, role, created_at]

    # ── Categories ──
...